Calculate consumer group lag
Before this change, calls to list or describe consumer groups would
include fields for lag and logEndOffset in their results (per
partition), but the values of these fields would always be 0.

This change figures out all of the TopicPartitions for which lag needs
to be calculated, and fetches the latest offsets for these
partitions. Then the lag is calculated as the difference between the
consumer group offset on a partition, and the latest offset of the
partition.