Couple all group info together in getGroupList
While we have to make two separate calls get get the
(group ID, description) and (topic partition, offset) maps, the latter
is per group, but doesn't contain information about the group in the
result.

This change adds a new ConsumerGroupInfo class (because Java doesn't
have useful things like tuples or algebraic data types or multiple
return (that "list" we were using where every two elements was a unit,
doesn't count)), which holds the information about a group together in
a way that's useful and makes it easier to operate on.