Catch odd wrapped exception, caught by test
ConsumerGroupsOAuthTestIT#testListWithInvalidToken caught an issue
where the server responds with a 503 if there's an invalid token.

For some reason, this SaslAuthenticationException is wrapped as
follows by the KafkaAdminClient:

```
new KafkaException("Failed to find brokers to send ListGroups",
    throwable);
```

which matches a different case further down in the
CommonHandler#processResponse method (line 123 at this commit).