Reduce complexity of method before adding condition
If I add another "else if" here without cleaning it up a bit,
checkstyle complains with "NCSS for this method is 101 (max allowed is
100). [JavaNCSS]"

It's possible that it could be cleaned up even further (see multiple
conditions which result in BAD_REQUEST), but the ordering might be
important because of the type hierarchy of the different Exceptions. I
haven't had time to check (I'd need to run all of the tests, and that
takes 40+ minutes) -- we can iterate on it in future.