Fix test flake and re-enable
The `RestOAuthTestIT#testListWithInvalidToken` test worked in
isolation, but would consistently fail when run as part of the entire
suite.

The reason it was failing in that scenario is because there's a static
global map in the AdminDeploymentManager to keep track of all of the
ports for admin server instances for tests, with the map keys being
the test method names. The problem in this case is that there were two
tests with the same name in two different classes, and the other one
would be run first. When it came time to run this test, there would
already be a port registered for the test method name from the other
test, and `ADMIN_PORTS.putIfAbsent` wouldn't "put" because "not
absent". This test would then have the wrong port registered, and the
call to list topics would time out (never reach the server because it
was trying on the wrong port).